+++ src/OpenTelemetry.Exporter.Prometheus/Implementation/PrometheusSerializerExt.cs
@@ -23,7 +23,10 @@ namespace OpenTelemetry.Exporter.Prometheus
     /// </summary>
     internal static partial class PrometheusSerializer
     {
-        private static readonly string[] MetricTypes = new string[] { "untyped", "counter", "gauge", "histogram", "summary" };
+        private static readonly string[] MetricTypes = new string[]
+        {
+            "untyped", "counter", "gauge", "summary", "histogram", "histogram", "histogram", "histogram", "untyped",
+        };
 
         public static int WriteMetric(byte[] buffer, int cursor, Metric metric)
         {
@@ -35,7 +38,7 @@ namespace OpenTelemetry.Exporter.Prometheus
             int metricType = (int)metric.MetricType >> 4;
             cursor = WriteTypeInfo(buffer, cursor, metric.Name, metric.Unit, MetricTypes[metricType]);
 
-            if (metric.MetricType != MetricType.Histogram)
+            if (!metric.MetricType.IsHistogram())
             {
                 foreach (ref var metricPoint in metric.GetMetricPoints())
                 {